b2: a malformed page cursor keeps the answer a Python slice gave it
The `version_history` cursor is a plaintext decimal offset frozen by M2's C6,
and the old path fed it straight to `rows[offset:offset + limit]` on a Python
list — so a nonsensical cursor produced a nonsensical but *defined* page. A
`usize` offset on the binding would have turned `cursor="-5"` into an
`OverflowError` instead, which is a contract change nobody asked for.

`version_page` therefore takes a signed offset and `py_slice` reproduces
CPython's `PySlice_AdjustIndices` exactly: each endpoint adjusted by the
length only if it is negative, then clamped independently, with a stop below
the start meaning an empty page. `range(10)[-5:3]` is empty on both sides of
this change, and the table in `py_slice_is_cpython_slice_arithmetic` was
checked against CPython's own numbers rather than reasoned about.

+++ crates/tgms-engine-py/src/lib.rs
@@ -870,8 +870,8 @@ impl NativeStore {
         t_b: i64,
         belief: &str,
         rel_types: Option<Vec<String>>,
-        offset: usize,
-        limit: usize,
+        offset: i64,
+        limit: i64,
     ) -> Res<Bound<'py, PyDict>> {
         let k = kind_of(kind)?;
         let belief = match belief {
